Add a folder for the option dtos
