+++ PollsAppBlazor/Client/Pages/Polls/Create.razor
@@ -1,5 +1,6 @@
 ﻿@page "/polls/create"
 @using Microsoft.AspNetCore.Authorization
+@using PollsAppBlazor.Shared.Options
 @using PollsAppBlazor.Shared.Polls
 @inject ISnackbar Snackbar
 @inject NavigationManager NavigationManager
