+++ PollsAppBlazor/Client/Pages/Polls/Poll.razor
@@ -1,5 +1,6 @@
 ﻿@page "/polls/{pollId:int}"
 @using Microsoft.AspNetCore.Authorization
+@using PollsAppBlazor.Shared.Options
 @using PollsAppBlazor.Shared.Polls
 @inject PublicClient PublicClient
 @inject NavigationManager NavigationManager
