+++ PollsAppBlazor/Client/Shared/PollForm.razor
@@ -1,4 +1,5 @@
-﻿@using PollsAppBlazor.Shared.Polls
+﻿@using PollsAppBlazor.Shared.Options
+@using PollsAppBlazor.Shared.Polls
 @inject NavigationManager NavigationManager
 @inject ISnackbar Snackbar
