+++ PollsAppBlazor/Shared/Polls/PollViewDto.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using PollsAppBlazor.Shared.Options;
+using System.Text.Json.Serialization;
 
 namespace PollsAppBlazor.Shared.Polls;
